refactor: Move to entity event
